+++ backend/routes/Auth/logout-route.js
@@ -1,0 +1,1 @@
+
